Paper 4 draft + 268 experiments + state of research synthesis
Paper 4: "Two Kinds of Not Knowing Yourself" — 5,722 words, 9 sections.
New findings integrated: RLHF spectral invisibility (§2.4), shape vs
direction dissociation (§2.5), direction-level anti-suppressant mechanism.

State of research synthesis covers all 268 experiments, 78+ findings,
16+ models, 2 retractions. Comprehensive as of June 8, 2026.

Includes experiments from phases 5-9 and results JSON files.
Two oversized results files (>100MB) excluded from GitHub.

+++ experiments/exp_convergence_v2.py
@@ -72,7 +72,7 @@ def load_model(model_name):
     if tokenizer.pad_token is None:
         tokenizer.pad_token = tokenizer.eos_token
     model = AutoModelForCausalLM.from_pretrained(
-        model_name, torch_dtype=torch.float16, device_map="auto",
+        model_name, torch_dtype=torch.bfloat16, device_map="auto",
         trust_remote_code=True, attn_implementation="eager"
     )
     model.eval()
